Add a prompt for limit on launch to templates.
Bumped AWX-Management version to 1.4.

Performing a self-update can be triggered by leaving one or both prompts
blanks when launching 'AWX-Template-Management'.

Prompting for limit allows template/workflows to be launched against a
single host in an inventory.  The host IP can be passed from a script as
an argument when launching a template/workflow.

The inventory host for the AWX Server no longer stores the
AWX-Management version number in the host variables.  This is now stored
as a fact that can be viewed in the facts tab for the AWX Server host.
This fixes an issue when scheduling the 'AWX-Template-Management'
template where the scheduler would not use the updated version number
causing the scheduled task to only perform a self-update.

The AWX-Management version number is now added to each template during
the create/update task instead of reading the version number in each
playbook.  Template updates are skipped if the template and server
versions match.